move resolver from main to individual package
