+++ cmd/falco/main.go
@@ -12,6 +12,7 @@ import (
 	"github.com/kyokomi/emoji"
 	"github.com/mattn/go-colorable"
 	"github.com/pkg/errors"
+	"github.com/ysugimoto/falco/resolver"
 	"github.com/ysugimoto/falco/terraform"
 )
 
@@ -139,20 +140,20 @@ func main() {
 
 	var fetcher Fetcher
 	// falco could lint multiple services so resolver should be a slice
-	var resolvers []Resolver
+	var resolvers []resolver.Resolver
 	switch fs.Arg(0) {
 	case subcommandTerraform:
 		fastlyServices, err := ParseStdin()
 		if err == nil {
-			resolvers = NewTerraformResolver(fastlyServices)
+			resolvers = resolver.NewTerraformResolver(fastlyServices)
 			fetcher = terraform.NewTerraformFetcher(fastlyServices)
 		}
 	case subcommandLint:
 		// "lint" command provides single file of service, then resolvers size is always 1
-		resolvers, err = NewFileResolvers(fs.Arg(1), c)
+		resolvers, err = resolver.NewFileResolvers(fs.Arg(1), c.IncludePaths)
 	default:
 		// "lint" command provides single file of service, then resolvers size is always 1
-		resolvers, err = NewFileResolvers(fs.Arg(0), c)
+		resolvers, err = resolver.NewFileResolvers(fs.Arg(0), c.IncludePaths)
 	}
 
 	if err != nil {
