+++ resolver/resolver.go
@@ -1,4 +1,4 @@
-package main
+package resolver
 
 import (
 	"github.com/pkg/errors"
